refactor(inventory): move inventory cmd files in inventory dir
